refactor(apple): spacchetta il workspace pazienti in viste coese
Prerequisito DS-2 della migrazione Lume nativa: separa il workspace pazienti in viste coese senza cambiare comportamento o resa.
